+++ vite.config.js
@@ -7,6 +7,27 @@ import VueI18nPlugin from "@intlify/unplugin-vue-i18n/vite"
 
 import svgLoader from "./src/svg-loader.ts"
 
+// https://v5.vite.dev/guide/api-plugin.html#transformindexhtml
+function customCSSPlugin() {
+  return {
+    name: "custom-css",
+    transformIndexHtml: {
+      order: "post",
+      handler(html) {
+        return [{
+          tag: "link",
+          attrs: {
+            // Load custom stylesheet
+            rel: "stylesheet",
+            href: "/assets/custom.css",
+          },
+          injectTo: "head",
+        }]
+      },
+    },
+  }
+}
+
 // https://vitejs.dev/config/
 export default ({ mode }) => {
   // import.meta.env is only available in application code
@@ -20,6 +41,7 @@ export default ({ mode }) => {
       VueI18nPlugin({
         include: path.resolve(path.dirname(fileURLToPath(import.meta.url)), "@/src/locales/**"),
       }),
+      customCSSPlugin(),
     ],
     resolve: {
       alias: {